requirements.txt
tensorflow>=2.13.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0